classes, conf: added bbclass, config bits for caync-based updates.
Added image-casync.bbclass which, when inherited, creates/populates
an image-specific casync chunks store with the image rootfs and
generates an image-version-specific casync index for reproducing an
identical copy of the rootfs.

Added casync-config.inc, which contain the global configuration bits
for image-casync.bbclass. Also added edgeos-enable-casync.int which
is an edgeos-specific configuration fragment for enabling casyn-
based updates for edgeos.