WiP: meson.bbclass: kludge to make sure we use our meson.bbclass.
Renamed our copy of meson.bbclass, a pathced copy of the one in
meta-oe, to casync_meson.bbclass and changed our casync recipe to
use that one.

Once the native rpath-related problems in the stock meta-oe class
are fixed we can drop our copy and change the casync recipe back
to use the stock one.

Signed-off-by: Krisztian Litkey <kli@iki.fi>